Assistant/requirements.txt
cryptography==37.0.4
geocoder==1.38.1
google_search_results==2.4.1
openai==0.25.0
pafy==0.5.5
pandas==1.4.2
playsound==1.3.0
pvporcupine==2.1.3
Py_Dictionary==4.1.4
PyAudio==0.2.12
pydictionary==2.0.1
pyjokes==0.6.0
pyowm==3.3.0
PySimpleGUI==4.60.4
python_vlc==3.0.16120
pyttsx3==2.90
requests==2.28.0
scikit_learn==1.2.0
SpeechRecognition==3.8.1
tinytuya==1.6.6
translate==3.6.1
